Using f64 for objective value

+++ src/lib.rs
@@ -126,7 +126,7 @@ impl fmt::Display for SparseLEIntegerPolyhedron<'_> {
 #[derive(Debug)]
 pub struct Solution {
     pub status: Status,
-    pub objective: i64,
+    pub objective: f64,
     pub solution: HashMap<String, i32>,
     pub error: Option<String>,
 }
@@ -294,7 +294,7 @@ pub fn solve_ilps<'a>(polytope: &mut SparseLEIntegerPolyhedron<'a>, objectives: 
         for obj in objectives.iter() {
 
             // Setup empty solution
-            let mut solution = Solution { status: Status::Undefined, objective: 0, solution: HashMap::new(), error: None };
+            let mut solution = Solution { status: Status::Undefined, objective: 0.0, solution: HashMap::new(), error: None };
 
             // Update the objective function
             for (j, v) in polytope.variables.iter().enumerate() {
@@ -337,7 +337,7 @@ pub fn solve_ilps<'a>(polytope: &mut SparseLEIntegerPolyhedron<'a>, objectives: 
                 },
                 2 => {
                     solution.status = Status::Feasible;
-                    solution.objective = glpk::glp_mip_obj_val(lp) as i64;
+                    solution.objective = glpk::glp_mip_obj_val(lp);
                     for (j, var) in polytope.variables.iter().enumerate() {
                         let x = glpk::glp_mip_col_val(lp, (j + 1) as i32);
                         solution.solution.insert(var.id.to_string(), x as i32);
@@ -353,7 +353,7 @@ pub fn solve_ilps<'a>(polytope: &mut SparseLEIntegerPolyhedron<'a>, objectives: 
                 }
                 5 => {
                     solution.status = Status::Optimal;
-                    solution.objective = glpk::glp_mip_obj_val(lp) as i64;
+                    solution.objective = glpk::glp_mip_obj_val(lp);
                     for (j, var) in polytope.variables.iter().enumerate() {
                         let x = glpk::glp_mip_col_val(lp, (j + 1) as i32);
                         solution.solution.insert(var.id.to_string(), x as i32);
@@ -800,7 +800,7 @@ mod tests {
         assert_eq!(solutions.len(), 1);
         let solution = &solutions[0];
         assert_eq!(solution.status, Status::Optimal);
-        assert_eq!(solution.objective, 1);
+        assert_eq!(solution.objective, 1.0);
         assert_eq!(solution.solution.get("a"), Some(&1));
         assert_eq!(solution.solution.get("b"), Some(&0));
     }
